cleanup filenames
